src/test/java/com/leetcode/algorithm/medium/combinationsumiii/SolutionTest.java
package com.leetcode.algorithm.medium.combinationsumiii;

import org.junit.jupiter.api.Test;

import java.util.Arrays;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    assertEquals(Arrays.asList(Arrays.asList(1, 2, 4)),
        solution.combinationSum3(3, 7));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    assertEquals(Arrays.asList(Arrays.asList(1, 2, 6), Arrays.asList(1, 3, 5), Arrays.asList(2, 3, 4)),
        solution.combinationSum3(3, 9));
  }
}
